Add support for parameters for constructor in testing.

+++ testlib/test_with_contracts.py
@@ -11,11 +11,16 @@ class TestWithContracts():
         self.build_contracts_dir = self.tmp_path / Path('build') / Path('contracts')
         self.json_compiled_fixture = self.fixtures_dir / Path("HelloWorld.json")
         self.json_compiled_file = self.build_contracts_dir / Path('HelloWorld.json')
+        self.json2_compiled_fixture = self.fixtures_dir / Path("HelloParameters.json")
+        self.json2_compiled_file = self.build_contracts_dir / Path('HelloParameters.json')
 
         self.build_dir.mkdir()
         self.build_contracts_dir.mkdir()
 
     def teardown_method(self, method):
-        self.json_compiled_file.unlink()
+        if self.json_compiled_file.exists():
+            self.json_compiled_file.unlink()
+        if self.json2_compiled_file.exists():
+            self.json2_compiled_file.unlink()
         self.build_contracts_dir.rmdir()
         self.build_dir.rmdir()
